+++ app/src/main/java/com/example/get_ripped/data/model/SetEntry.kt
@@ -2,5 +2,5 @@ package com.example.get_ripped.data.model
 
 data class SetEntry(
     val reps: Int,
-    val weight: Int
+    val weight: Float
 )
